Update synonyms.py
Code of an intelligent system that can learn to answer questions like the typical multiple choice question asked in the TOEFL and SAT exams of which word is most similar in meaning to another given word. In order to do that, the system approximates the semantic similarity of any pair of words. The semantic similarity between two words is the measure of the closeness of their meanings. In order to answer the question, the program computes the semantic similarity between the word given and all the possible answers, and picks the answer with the highest semantic similarity to the given word. More precisely, given a word w and a list of potential synonyms s1, s2, s3, s4, the program computes the similarities of (w, s1), (w, s2), (w, s3), (w, s4) and chooses the word whose similarity to w is the highest. The program measures the semantic similarity of pairs of words by first computing a semantic descriptor vector of each of the words, and then taking the similarity measure to be the cosine similarity between
the two vectors. Given a text with n words denoted by (w1, w2, …, w(n)) and a word w, let descw be the semantic descriptor vector of w computed using the text. descw is an n-dimensional vector. The i-th coordinate of descw is the number of sentences in which both w and wi occur. For efficiency’s sake, the code stores the semantic
descriptor vector as a dictionary, not storing the zeros that correspond to words which don’t co-occur with w.