Create FK.py
Different functions that deal with processing text files in conjunction to dictionaries. There's a function that opens a text file and prints all the lines in the text that contain a given word. A function that converts a dictionary to a string with the key and value pairs separated by new lines. A function that creates a Python dictionary whose keys are words, and whose values are lists of phones based on the CMU Pronouncing dictionary that contains the list of all the vowel and consonant phones in the dictionary. Also a function that creates a Python dictionary whose keys are phone codes, and whose values are the categories (vowels, stops, affricates, fricatives, liquids, nasals, semivowels). e A function that takes in a word (assuming it is in the Pronouncing Dictionary) and the dictionaries made previously, and returns the number of vowels in the word. A function that approximates the number of syllables in a word. And lastly, a function that evaluates the Flesch-Kincaid Readability Grade level of a
text.